Add testcontainers support for Kafka
Lift the testcontainers containers for Kafka and the Confluent
schema registry into a new `safir.testing.containers` module and
document them so that they can be used by other packages to test
with a real Kafka server.

Extensively refactor and simplify the Kafka test container to remove
things we weren't using, update the defaults to match our use cases,
and simplify the internal class structure and public API.
