+++ changelog.d/20250522_182327_rra_DM_50964a.md
@@ -1,0 +1,3 @@
+### New features
+
+- Add `safir.testing.containers.FullKafkaContainer` and `safir.testing.containers.SchemaRegistryContainer` classes based on Testcontainers that create containers for Kafka and the Confluent schema registry, respectively. These classes are intended for use in fixtures or test configuration to enable testing against a local temporary Kafka cluster.
