+++ docs/api.rst
@@ -103,6 +103,9 @@ API reference
 .. automodapi:: safir.slack.webhook
    :include-all-objects:
 
+.. automodapi:: safir.testing.containers
+   :include-all-objects:
+
 .. automodapi:: safir.testing.gcs
    :include-all-objects:
