+++ docs/user-guide/kafka/index.rst
@@ -14,3 +14,4 @@ Guides
 
    kafka-connection-settings
    pydantic-schema-manager
+   testing
