+++ safir/src/safir/testing/containers/__init__.py
@@ -1,0 +1,9 @@
+"""Safir-provided test containers."""
+
+from ._kafka import FullKafkaContainer
+from ._schema_registry import SchemaRegistryContainer
+
+__all__ = [
+    "FullKafkaContainer",
+    "SchemaRegistryContainer",
+]
