+++ safir/src/safir/testing/containers/_constants.py
@@ -1,0 +1,10 @@
+"""Constants for test containers."""
+
+CONFLUENT_VERSION_TAG = "7.6.0"
+"""Docker image tag for Confluent images (Kafka and schema registry)."""
+
+KAFKA_REPOSITORY = "confluentinc/cp-kafka"
+"""Docker repository for Kafka."""
+
+SCHEMA_REGISTRY_REPOSITORY = "confluentinc/cp-schema-registry"
+"""Docker repository for the Confluent schema registry."""
