+++ safir/src/safir/testing/containers/data/generate-kafka-secrets.bash
